# server/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dashboard Restaurante API",
    "version": "1.0.0",
    "description": "API REST para gestión de pedidos, reservas, ventas, inventario y clientes de restaurante.",
    "contact": { "email": "soporte@mirestaurante.com" }
  },
  "servers": [
    { "url": "http://localhost:9000/api", "description": "Desarrollo local" },
    { "url": "https://tu-dominio.com/api", "description": "Producción" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Producto": {
        "type": "object",
        "properties": {
          "id":           { "type": "integer" },
          "nombre":       { "type": "string" },
          "precio":       { "type": "number" },
          "stock":        { "type": "integer" },
          "categoria_id": { "type": "integer" },
          "activo":       { "type": "boolean" }
        }
      },
...
      "Cliente": {
        "type": "object",
        "properties": {
          "id":           { "type": "integer" },
          "nombre":       { "type": "string" },
          "email":        { "type": "string" },
          "telefono":     { "type": "string" },
          "rut":          { "type": "string" },
          "tipo_cliente": { "type": "string", "enum": ["persona","empresa"] },
          "estado":       { "type": "string", "enum": ["Nuevo","Regular","VIP"] },
          "visitas":      { "type": "integer" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/clientes": {
      "get": {
        "tags": ["Clientes"],
        "summary": "Listar clientes",
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "Búsqueda por nombre, email o RUT" }
        ],
        "responses": {
          "200": { "description": "Lista de clientes", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Cliente" } } } } }
        }
      },
      "post": {
        "tags": ["Clientes"],
        "summary": "Crear cliente",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nombre"],
                "properties": {
                  "nombre":       { "type": "string" },
                  "email":        { "type": "string" },
                  "telefono":     { "type": "string" },
                  "rut":          { "type": "string" },
                  "tipo_cliente": { "type": "string", "enum": ["persona","empresa"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Cliente creado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Cliente" } } } }
        }
      }
    },
    "/clientes/{id}": {
      "get": {
        "tags": ["Clientes"],
        "summary": "Obtener cliente con historial de pedidos",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Cliente con pedidos" } }
      },
      "put": {
        "tags": ["Clientes"],
        "summary": "Actualizar cliente",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Cliente" } } }
        },
        "responses": { "200": { "description": "Cliente actualizado" } }
      },
      "delete": {
        "tags": ["Clientes"],
        "summary": "Eliminar cliente (admin)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Cliente eliminado" } }
      }
    },
...
    "/config": {
      "get": {
        "tags": ["Configuración"],
        "summary": "Obtener configuración del negocio",
        "description": "Retorna nombre, moneda, impuesto, prefijo de ticket, logo, etc.",
        "responses": { "200": { "description": "Configuración actual del restaurante" } }
      },
      "put": {
        "tags": ["Configuración"],
        "summary": "Guardar configuración del negocio",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre_negocio":  { "type": "string" },
                  "moneda":          { "type": "string", "example": "CLP" },
                  "currency":        { "type": "string", "example": "$" },
                  "tax_rate":        { "type": "number", "example": 19 },
                  "impuesto_activo": { "type": "boolean" },
                  "prefijo_ticket":  { "type": "string", "example": "TKT" },
                  "numero_inicial":  { "type": "integer", "example": 1 }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Configuración guardada" } }
      }
    },
    "/config/logo": {
      "post": {
        "tags": ["Configuración"],
        "summary": "Subir logo del restaurante",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": { "type": "object", "properties": { "logo": { "type": "string", "format": "binary" } } }
            }
          }
        },
        "responses": { "200": { "description": "Logo subido. Retorna `logo_url`." } }
      }
    },
    "/productos": {
      "get": {
        "tags": ["Productos"],
        "summary": "Listar productos",
        "parameters": [
          { "name": "categoria_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "activo", "in": "query", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": { "description": "Lista de productos", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Producto" } } } } }
        }
      },
      "post": {
        "tags": ["Productos"],
        "summary": "Crear producto",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nombre", "precio"],
                "properties": {
                  "nombre":       { "type": "string" },
                  "precio":       { "type": "number" },
                  "stock":        { "type": "integer" },
                  "categoria_id": { "type": "integer" },
                  "descripcion":  { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Producto creado" } }
      }
    },
    "/productos/{id}": {
      "get": {
        "tags": ["Productos"],
        "summary": "Obtener producto por ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Producto encontrado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Producto" } } } },
          "404": { "description": "Producto no encontrado" }
        }
      },
      "put": {
        "tags": ["Productos"],
        "summary": "Actualizar producto",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Producto" } } }
        },
        "responses": { "200": { "description": "Producto actualizado" } }
      },
      "delete": {
        "tags": ["Productos"],
        "summary": "Eliminar producto (soft delete, marca como inactivo)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Producto desactivado" } }
      }
    },
...
    "/caja/abrir": {
      "post": {
        "tags": ["Caja"],
        "summary": "Abrir caja",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "fondo_inicial": { "type": "number" } } }
            }
          }
        },
        "responses": { "200": { "description": "Caja abierta" } }
      }
    },
    "/caja/cerrar": {
      "post": {
        "tags": ["Caja"],
        "summary": "Cerrar caja con arqueo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "monto_contado": { "type": "number" }, "notas": { "type": "string" } } }
            }
          }
        },
        "responses": { "200": { "description": "Caja cerrada con resumen" } }
      }
    },
...
  },
...
}